fixed cleandb.ts

+++ server/src/seeds/cleanDb.ts
@@ -1,11 +1,33 @@
+// import models from '../models/index.js';
+// import db from '../config/connection.js';
+
+// export default async (modelName: "Question", collectionName: string) => {
+//   try {
+//     let modelExists = await models[modelName].db.db.listCollections({
+//       name: collectionName
+//     }).toArray()
+
+//     if (modelExists.length) {
+//       await db.dropCollection(collectionName);
+//     }
+//   } catch (err) {
+//     throw err;
+//   }
+// }
+//Fixed?
 import models from '../models/index.js';
 import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
+    const model = models[modelName];
+    if (!model || !model.db || !model.db.db) {
+      throw new Error(`Model or database not found for modelName: ${modelName}`);
+    }
+
+    let modelExists = await model.db.db.listCollections({
       name: collectionName
-    }).toArray()
+    }).toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
